+++ yandex/data_source_yandex_mdb_postgresql_cluster_test.go
@@ -167,10 +167,6 @@ func testAccDataSourceMDBPGClusterAttributesCheck(datasourceName string, resourc
 				"folder_id",
 			},
 			{
-				"health",
-				"health",
-			},
-			{
 				"host.#",
 				"host.#",
 			},